pyproject.toml
[project]
name = "interior-visual-search"
version = "0.1.0"
requires-python = ">=3.11,<3.12"
dependencies = [
  "fastapi==0.128.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.3.0",
  "python-multipart>=0.0.9",
  "orjson>=3.10.0",
  "loguru>=0.7.2",

  "pinecone==8.0.0",

  # Image / ML
  "pillow>=10.4.0",
  "numpy>=2.0.0",
  "opencv-python>=4.10.0.84",

  # Optional (if you actually run deep models in-process)
  "torch>=2.4.0",
  "torchvision>=0.19.0",
  "transformers>=4.44.0",
  "hf-xet",
  "ultralytics>=8.3.0",  # Includes SAM2.1 support
  
  # Batch ingestion dependencies
  "pandas>=2.0.0",
  "openpyxl>=3.1.0",
  "httpx>=0.27.0",
  "tqdm>=4.66.0",
]

[tool.uv]
dev-dependencies = [
  "pytest>=8.2.0",
  "httpx>=0.27.0",
  "ruff>=0.5.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
testpaths = ["app/tests"]